requirements.txt
certifi==2018.1.18
chardet==3.0.4
idna==2.6
pyodbc==4.0.22
python-dateutil==2.7.2
python-tds==1.9.1
pytz==2018.3
PyYAML==3.12
requests==2.18.4
six==1.11.0
tzlocal==1.5.1
urllib3==1.22
